Reorganize and refactor
