+++ src/app/events/[slug]/components/ScheduleSection/Participants.tsx
@@ -2,14 +2,17 @@ import React from 'react'
 
 import type { NonEmptyArray } from '@/types/utils'
 
-import type { Participant } from '../../schemas/ScheduleSchema'
+import type { Participant } from '../../../schemas/ScheduleSchema'
 
 export type ParticipantsProps = {
   title: string
   participants: NonEmptyArray<Participant>
 }
 
-export function Participants({ title, participants }: ParticipantsProps) {
+export function ScheduleParticipants({
+  title,
+  participants,
+}: ParticipantsProps) {
   return (
     <p className="text-sm">
       <span className="font-semibold">{title}:</span>{' '}
